+++ src/main/java/com/formulasearchengine/mathmltools/mml/CMMLInfo.java
@@ -26,28 +26,15 @@ import net.sf.saxon.s9api.SaxonApiException;
 import net.sf.saxon.s9api.XQueryExecutable;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.w3c.dom.Attr;
-import org.w3c.dom.CDATASection;
-import org.w3c.dom.Comment;
-import org.w3c.dom.DOMConfiguration;
 import org.w3c.dom.DOMException;
-import org.w3c.dom.DOMImplementation;
 import org.w3c.dom.Document;
-import org.w3c.dom.DocumentFragment;
-import org.w3c.dom.DocumentType;
-import org.w3c.dom.Element;
-import org.w3c.dom.EntityReference;
-import org.w3c.dom.NamedNodeMap;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
-import org.w3c.dom.ProcessingInstruction;
-import org.w3c.dom.Text;
-import org.w3c.dom.UserDataHandler;
 
 /**
  * @author Moritz Schubotz
  */
-public class CMMLInfo implements Document {
+public class CMMLInfo extends CMMLInfoBase implements Document {
 
     //For XML math processing
     public static final String NS_MATHML = "http://www.w3.org/1998/Math/MathML";
@@ -71,7 +58,7 @@ public class CMMLInfo implements Document {
             "equivalent"
     );
 
-    private Document cmmlDoc;
+
     private XQueryGenerator queryGenerator = null;
     private XQueryExecutable xQueryExecutable;
     private boolean isStrict;
@@ -105,6 +92,10 @@ public class CMMLInfo implements Document {
 
     public CMMLInfo(Document document, boolean preserveAnnotations) {
         constructor(document, false, preserveAnnotations);
+    }
+
+    public CMMLInfo(CMMLInfoBase cmmlInfoBase) {
+        new CMMLInfoBase();
     }
 
     public static CMMLInfo newFromSnippet(String snippet) throws IOException, ParserConfigurationException {
@@ -210,32 +201,6 @@ public class CMMLInfo implements Document {
         }
     }
 
-    @Override
-    public final Node adoptNode(Node node) throws DOMException {
-        return cmmlDoc.adoptNode(node);
-    }
-
-    @Override
-    public final Node appendChild(Node node) throws DOMException {
-        return cmmlDoc.appendChild(node);
-    }
-
-
-    @SuppressWarnings("superClone")
-    @Override
-    public final CMMLInfo clone() {
-        return new CMMLInfo(this);
-    }
-
-    @Override
-    public final Node cloneNode(boolean b) {
-        return cmmlDoc.cloneNode(b);
-    }
-
-    @Override
-    public final short compareDocumentPosition(Node node) throws DOMException {
-        return cmmlDoc.compareDocumentPosition(node);
-    }
 
     private void constructor(Document cmml, Boolean fixNamespace, Boolean preserveAnnotations) {
         cmmlDoc = cmml;
@@ -247,56 +212,6 @@ public class CMMLInfo implements Document {
             removeAnnotations();
         }
         removeElementsByName("id");
-    }
-
-    @Override
-    public final Attr createAttribute(String s) throws DOMException {
-        return cmmlDoc.createAttribute(s);
-    }
-
-    @Override
-    public final Attr createAttributeNS(String s, String s1) throws DOMException {
-        return cmmlDoc.createAttributeNS(s, s1);
-    }
-
-    @Override
-    public final CDATASection createCDATASection(String s) throws DOMException {
-        return cmmlDoc.createCDATASection(s);
-    }
-
-    @Override
-    public final Comment createComment(String s) {
-        return cmmlDoc.createComment(s);
-    }
-
-    @Override
-    public final DocumentFragment createDocumentFragment() {
-        return cmmlDoc.createDocumentFragment();
-    }
-
-    @Override
-    public final Element createElement(String s) throws DOMException {
-        return cmmlDoc.createElement(s);
-    }
-
-    @Override
-    public final Element createElementNS(String s, String s1) throws DOMException {
-        return cmmlDoc.createElementNS(s, s1);
-    }
-
-    @Override
-    public final EntityReference createEntityReference(String s) throws DOMException {
-        return cmmlDoc.createEntityReference(s);
-    }
-
-    @Override
-    public final ProcessingInstruction createProcessingInstruction(String s, String s1) throws DOMException {
-        return cmmlDoc.createProcessingInstruction(s, s1);
-    }
-
-    @Override
-    public final Text createTextNode(String s) {
-        return cmmlDoc.createTextNode(s);
     }
 
     private void fixNamespaces() {
@@ -326,21 +241,6 @@ public class CMMLInfo implements Document {
         } catch (final DOMException e) {
             //Ignore any errors thrown if element does not exist
         }
-    }
-
-    @Override
-    public final NamedNodeMap getAttributes() {
-        return cmmlDoc.getAttributes();
-    }
-
-    @Override
-    public final String getBaseURI() {
-        return cmmlDoc.getBaseURI();
-    }
-
-    @Override
-    public final NodeList getChildNodes() {
-        return cmmlDoc.getChildNodes();
     }
 
     public final Double getCoverage(Multiset<String> queryTokens) {
@@ -386,36 +286,6 @@ public class CMMLInfo implements Document {
         return cmmlDoc;
     }
 
-    @Override
-    public final DocumentType getDoctype() {
-        return cmmlDoc.getDoctype();
-    }
-
-    @Override
-    public final Element getDocumentElement() {
-        return cmmlDoc.getDocumentElement();
-    }
-
-    @Override
-    public final String getDocumentURI() {
-        return cmmlDoc.getDocumentURI();
-    }
-
-    @Override
-    public final void setDocumentURI(String s) {
-        cmmlDoc.setDocumentURI(s);
-    }
-
-    @Override
-    public final DOMConfiguration getDomConfig() {
-        return cmmlDoc.getDomConfig();
-    }
-
-    @Override
-    public final Element getElementById(String s) {
-        return cmmlDoc.getElementById(s);
-    }
-
     public final Multiset<String> getElements() {
         try {
             Multiset<String> list = HashMultiset.create();
@@ -442,126 +312,6 @@ public class CMMLInfo implements Document {
             }
         }
         return cachedElements;
-    }
-
-    @Override
-    public final NodeList getElementsByTagName(String s) {
-        return cmmlDoc.getElementsByTagName(s);
-    }
-
-    @Override
-    public final NodeList getElementsByTagNameNS(String s, String s1) {
-        return cmmlDoc.getElementsByTagNameNS(s, s1);
-    }
-
-    @Override
-    public final Object getFeature(String s, String s1) {
-        return cmmlDoc.getFeature(s, s1);
-    }
-
-    @Override
-    public final Node getFirstChild() {
-        return cmmlDoc.getFirstChild();
-    }
-
-    @Override
-    public final DOMImplementation getImplementation() {
-        return cmmlDoc.getImplementation();
-    }
-
-    @Override
-    public final String getInputEncoding() {
-        return cmmlDoc.getInputEncoding();
-    }
-
-    @Override
-    public final Node getLastChild() {
-        return cmmlDoc.getLastChild();
-    }
-
-    @Override
-    public final String getLocalName() {
-        return cmmlDoc.getLocalName();
-    }
-
-    @Override
-    public final String getNamespaceURI() {
-        return cmmlDoc.getNamespaceURI();
-    }
-
-    @Override
-    public final Node getNextSibling() {
-        return cmmlDoc.getNextSibling();
-    }
-
-    @Override
-    public final String getNodeName() {
-        return cmmlDoc.getNodeName();
-    }
-
-    @Override
-    public final short getNodeType() {
-        return cmmlDoc.getNodeType();
-    }
-
-    @Override
-    public final String getNodeValue() throws DOMException {
-        return cmmlDoc.getNodeValue();
-    }
-
-    @Override
-    public final void setNodeValue(String s) throws DOMException {
-        cmmlDoc.setNodeValue(s);
-    }
-
-    @Override
-    public final Document getOwnerDocument() {
-        return cmmlDoc.getOwnerDocument();
-    }
-
-    @Override
-    public final Node getParentNode() {
-        return cmmlDoc.getParentNode();
-    }
-
-    @Override
-    public final String getPrefix() {
-        return cmmlDoc.getPrefix();
-    }
-
-    @Override
-    public final void setPrefix(String s) throws DOMException {
-        cmmlDoc.setPrefix(s);
-    }
-
-    @Override
-    public final Node getPreviousSibling() {
-        return cmmlDoc.getPreviousSibling();
-    }
-
-    @Override
-    public final boolean getStrictErrorChecking() {
-        return cmmlDoc.getStrictErrorChecking();
-    }
-
-    @Override
-    public final void setStrictErrorChecking(boolean b) {
-        cmmlDoc.setStrictErrorChecking(b);
-    }
-
-    @Override
-    public final String getTextContent() throws DOMException {
-        return cmmlDoc.getTextContent();
-    }
-
-    @Override
-    public final void setTextContent(String s) throws DOMException {
-        cmmlDoc.setTextContent(s);
-    }
-
-    @Override
-    public final Object getUserData(String s) {
-        return cmmlDoc.getUserData(s);
     }
 
     public final XQueryExecutable getXQuery(Boolean useCache) {
@@ -599,61 +349,6 @@ public class CMMLInfo implements Document {
 
     public void setQueryGenerator(XQueryGenerator queryGenerator) {
         this.queryGenerator = queryGenerator;
-    }
-
-    @Override
-    public final String getXmlEncoding() {
-        return cmmlDoc.getXmlEncoding();
-    }
-
-    @Override
-    public final boolean getXmlStandalone() {
-        return cmmlDoc.getXmlStandalone();
-    }
-
-    @Override
-    public final void setXmlStandalone(boolean b) throws DOMException {
-        cmmlDoc.setXmlStandalone(b);
-    }
-
-    @Override
-    public final String getXmlVersion() {
-        return cmmlDoc.getXmlVersion();
-    }
-
-    @Override
-    public final void setXmlVersion(String s) throws DOMException {
-        cmmlDoc.setXmlVersion(s);
-    }
-
-    @Override
-    public final boolean hasAttributes() {
-        return cmmlDoc.hasAttributes();
-    }
-
-    @Override
-    public final boolean hasChildNodes() {
-        return cmmlDoc.hasChildNodes();
-    }
-
-    @Override
-    public final Node importNode(Node node, boolean b) throws DOMException {
-        return cmmlDoc.importNode(node, b);
-    }
-
-    @Override
-    public final Node insertBefore(Node node, Node node1) throws DOMException {
-        return cmmlDoc.insertBefore(node, node1);
-    }
-
-    @Override
-    public final boolean isDefaultNamespace(String s) {
-        return cmmlDoc.isDefaultNamespace(s);
-    }
-
-    @Override
-    public final boolean isEqualNode(Node node) {
-        return cmmlDoc.isEqualNode(node);
     }
 
     public final boolean isEquation(boolean useCache) throws XPathExpressionException {
@@ -694,44 +389,9 @@ public class CMMLInfo implements Document {
         return null;
     }
 
-    @Override
-    public final boolean isSameNode(Node node) {
-        return cmmlDoc.isSameNode(node);
-    }
-
-    @Override
-    public final boolean isSupported(String s, String s1) {
-        return cmmlDoc.isSupported(s, s1);
-    }
-
-    @Override
-    public final String lookupNamespaceURI(String s) {
-        return cmmlDoc.lookupNamespaceURI(s);
-    }
-
-    @Override
-    public final String lookupPrefix(String s) {
-        return cmmlDoc.lookupPrefix(s);
-    }
-
-    @Override
-    public final void normalize() {
-        cmmlDoc.normalize();
-    }
-
-    @Override
-    public final void normalizeDocument() {
-        cmmlDoc.normalizeDocument();
-    }
-
     private void removeAnnotations() {
         removeElementsByName("annotation");
         removeElementsByName("annotation-xml");
-    }
-
-    @Override
-    public final Node removeChild(Node node) throws DOMException {
-        return cmmlDoc.removeChild(node);
     }
 
     private void removeElementsByName(String name) {
@@ -742,21 +402,6 @@ public class CMMLInfo implements Document {
                 node.getParentNode().removeChild(node);
             }
         }
-    }
-
-    @Override
-    public final Node renameNode(Node node, String s, String s1) throws DOMException {
-        return cmmlDoc.renameNode(node, s, s1);
-    }
-
-    @Override
-    public final Node replaceChild(Node node, Node node1) throws DOMException {
-        return cmmlDoc.replaceChild(node, node1);
-    }
-
-    @Override
-    public final Object setUserData(String s, Object o, UserDataHandler userDataHandler) {
-        return cmmlDoc.setUserData(s, o, userDataHandler);
     }
 
     public final CMMLInfo toDataCmml() {
