+++ src/test/java/com/formulasearchengine/mathmltools/mml/TreeWriterTest.java
@@ -1,10 +1,16 @@
 package com.formulasearchengine.mathmltools.mml;
 
 import javax.xml.parsers.ParserConfigurationException;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.number.OrderingComparison.greaterThan;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertThat;
 
 import java.io.IOException;
+import java.util.AbstractMap;
+import java.util.ArrayList;
 
+import javax.xml.xpath.XPathExpressionException;
 import org.junit.Test;
 
 public class TreeWriterTest {
